+++ src/main/java/io/bytom/api/UnconfirmedTransaction.java
@@ -112,7 +112,7 @@ public class UnconfirmedTransaction {
         /**
          * The number of units of the asset being issued or spent.
          */
-        private Integer amount;
+        private long amount;
 
         /**
          * The definition of the asset being issued or spent (possibly null).
@@ -156,7 +156,7 @@ public class UnconfirmedTransaction {
         /**
          * The number of units of the asset being controlled.
          */
-        private Integer amount;
+        private long amount;
 
         /**
          * The definition of the asset being controlled (possibly null).
